feat: Enable setting history limits for the Kubernetes CronJob
By default, a Kubernetes cluster keeps a record of the 3 most recent
successful and the 1 most recent failed instance of a CronJob.

For the retirement pipeline (which normally runs daily) it may be wise
to retain a record of maybe the last 7, or even the last 30 runs.

Thus, introduce two new configuration parameters:

* RETIREMENT_K8S_CRONJOB_HISTORYLIMIT_FAILURE (default 1)
* RETIREMENT_K8S_CRONJOB_HISTORYLIMIT_SUCCESS (default 3)

... and set the CronJob's failedJobsHistoryLimit and
successfulJobsHistoryLimit properties accordingly.

Reference:
https://kubernetes.io/docs/tasks/job/automated-tasks-with-cron-jobs/