fix: Bump the Tutor version dependency to >=16.1.4
We rely on the Tutor OPENEDX_COMMON_VERSION variable to determine
the tag (or branch) to fetch for the Tubular repository.

In Tutor 16.1.4, the default value for OPENEDX_COMMON_VERSION changed
from open-release/palm.2 to open-release/palm.3. This means we now
need a version bump in our own install_requires list, so that we
require Tutor>=16.1.4.

Since this breaks compatibility with Open edX Olive (and will thus
require a major release bump on our part), update the README
accordingly.

Reference:
https://github.com/overhangio/tutor/pull/923