fix: Bump the Tutor version dependency to >=16.1.7
We rely on the Tutor OPENEDX_COMMON_VERSION variable to determine
the tag (or branch) to fetch for the Tubular repository.

In Tutor 16.1.6, the default value for OPENEDX_COMMON_VERSION changed
from open-release/palm.3 to open-release/palm.4. This means we now
need a version bump in our own install_requires list, so that we
require Tutor>=16.1.6.

However, the Tutor 16.1.6 package as published to PyPI contains a bug
that makes it uninstallable, so we have to require Tutor>=16.1.7
instead.

References:
https://github.com/overhangio/tutor/releases/tag/v16.1.6
https://github.com/overhangio/tutor/issues/943