fix: Drop support for retiring E-Commerce Service accounts
This never worked, and can't ever work, due to limitations in the
E-Commerce service itself.

Thus, rather than rely on whether ECOMMERCE_HOST is defined to decide
whether the retirement pipeline for that service should be enabled,
always disable it.

Also, update the README explaining that this plugin can't be used to
retire E-Commerce accounts.

Fixes #36.